Modify HorizontalAlign tremendously.
Now instead of calculating an angle, it simply uses the horizontal offset as
input for the PID loop.

Note that there is debug stuff in this commit and there has yet to be a heart
beat (something established to make sure that the roborio and pi are synced on
numbers) applied.