Add proposed changes.
One proposal is to cap the output of PID to not pass 360 degrees for the
turntable. Horizontal align in the camera tracking bit is adjusted for this
because the camera might be wondering why the turntable is wandering away from
the goal.

This cap means that the turn table needs to be zeroed so that the turn table
can only move in the counter clockwise direction.

Another proposal is to use turn table on target as a subsystem method instead
of just calling the pid controller object. :tongue:

I've updated all the commands that were affected by the above change.